OpenGL now replaces PDL::Graphics::OpenGL!
With a bit of hardwiring, this version of PDL builds on
cygwin using the POGL module rather than the internal PP
one (PDL::Graphics::OpenGL).  All the 3d demos run and
the Tk3d does as well although it crashes the cygwin X
server.  Need to check on additional platforms....  The

3dtest.pl in the top directory is a simple imagrgb test
program for debugging.