Merge in use of POGL for basic TriD functions
These changes allow one to toggle between the legacy TriD
PP OpenGL bindings and the OpenGL module bindings (a.k.a. POGL)
by setting the USE_POGL option appropriately in perldl.conf.

After some testing of the new code on various platforms, the
final stage of the refactoring is to replace the X11 + GLX
event and window code with GLUT based code.