Fix problem with extra glClear() calls
This resulted in blank displays on systems with acceleration,
particularly ones without X11 at all.  The origin of the problem
appears to the the fact that the display() routine being called
by the TriD graphics is not callable as a subroutine which is
what glutDisplayFunc() actually needs.  This partial fix to
display just returns immediately if called with no arguments.
It is hoped that a more complete solution to the problem can
wait for a more thorough refactoring to use GLUT functionality
directly rather than mimic the X11 interface.