Add Stefan Evert patch to twiddle for Mac OS X
Apparently calling glutMainLoopEvent() every twiddle loop
caused resize events not to be correctly processed on Mac
OS X with Apple GLUT (not FreeGLUT).  This patch only
calls glutMainLoopEvent() if there are no unprocessed
events in the fake XEvent queue.  This problem and the
confusion may eventually go away if we move away from
the pseudo-X11 event processing to a native GLUT callback
framework.