Remove deprecated TriD OpenGL build code
The legacy X11+PP build of TriD OpenGL bindings has been
removed.  You'll need to use the new Perl OpenGL based
build which has proven to be more portable and reliable.
This prepares TriD graphics development for the next
stage of refactoring for support and functionality.

The t/opengl.t was updated not to run if the TriD graphics
had not been built.