Graphics/TriD/TriD/Objects.pm: Add a table of contents
That module contains several packages, but no own code.
The patch adds a list of contained packages and their
object hierarchy to the module's POD.

I also removed the reference to the dead Tk demo.

Adding at least one line of description for each of the
packages is yet to be done.